moved calc_chi2 to fitting
