+++ txpipe/utils/fitting.py
@@ -33,4 +33,42 @@ def fit_straight_line(x, y, y_err=None):
 
 def line(slp,vrbl,intcpt):
     return slp * vrbl + intcpt
+
+def calc_chi2(y, err, yfit , v = False):
+    """
+    Compute chi2 between data and fitted curve
+
+    Parameters
+    ----------
+    y: array
+        y values of data
+    err: array (1D or 2D)
+        either error bars (if independent data points)
+        or covariance matrix
+    yfit: array
+        fitted values of data
+    v: bool
+        verbose output
+
+    Returns
+    -------
+    chi2: float
+    """
+    if err.shape == (len(y),len(y)):
+        #use full covariance
+        if v:
+            print('cov_mat chi2')
+        inv_cov = np.linalg.inv( np.matrix(err) )
+        chi2 = 0
+        for i in range(len(y)):
+            for j in range(len(y)):
+                chi2 = chi2 + (y[i]-yfit[i])*inv_cov[i,j]*(y[j]-yfit[j])
+        return chi2
+        
+    elif err.shape == (len(y),):
+        if v:
+            print('diagonal chi2')
+        return sum(((y-yfit)**2.)/(err**2.))
+    else:
+        raise IOError('error in err or cov_mat input shape')
